refactor(daemon): extract the pure stability loop into a leaf (#1633 review)
The deferred-interaction-outcome owner keeps the seam, the pending-record
ownership, and the R7 clear; the quiet-window polling loop and the
baseline-distrust verdict move to post-gesture-stability.ts, parameterized
by hooks (capture, surface reader, the three signature comparators) so the
leaf imports no cycle owners and no SessionState — verified outside the R9
cycle, which stays at main's exact 76/20. Owner drops 539 -> 374 lines.

The no-effect corroboration keeps comparing against the ORIGINAL pre-gesture
baseline (never a mid-loop rebased one), now stated in the leaf's doc. The
stabilization loop suite drives the unchanged public adapter; the verdict
suite wires the real classifier through the leaf's hook.
