+++ src/daemon/__tests__/post-gesture-stabilization-verdict.test.ts
@@ -1,12 +1,27 @@
 import assert from 'node:assert/strict';
 import { test } from 'vitest';
-import { buildInteractionSurfaceSignature } from '../interaction-outcome-policy.ts';
-import { decidePostGestureStabilityVerdict } from '../deferred-interaction-outcome.ts';
+import {
+  buildInteractionSurfaceSignature,
+  classifyBaselineSurfaceEvidence,
+  type InteractionSurfaceSignature,
+} from '../interaction-outcome-policy.ts';
+import { decidePostGestureStabilityVerdict as decideWithHooks } from '../post-gesture-stability.ts';
 import {
   applicationRootNode,
   keyboardWindowNodes,
   pickupSnapshot,
 } from './post-gesture-stabilization-fixtures.ts';
+
+// The pure verdict takes its baseline comparator as a hook; every case below
+// wires the real classifier, so this file keeps testing the same
+// classify-then-decide composition shipping callers run.
+const decidePostGestureStabilityVerdict = (params: {
+  needsBaselineDistrust: boolean;
+  baselineSignature: InteractionSurfaceSignature | undefined;
+  quietSignature: InteractionSurfaceSignature;
+  elapsedMs: number;
+  distrustCapMs: number;
+}) => decideWithHooks({ ...params, classifyBaselineEvidence: classifyBaselineSurfaceEvidence });
 
 // ---------------------------------------------------------------------------
 // #1542 defect 2: baseline-comparison distrust.
