+++ src/daemon/deferred-interaction-outcome.ts
@@ -2,7 +2,6 @@ import type { SnapshotCaptureAnnotations } from '@agent-device/contracts/capture
 import { isApplePlatform, isMobilePlatform } from '@agent-device/kernel/device';
 import type { SnapshotState } from '@agent-device/kernel/snapshot';
 import type { CommandFlags } from '../core/dispatch.ts';
-import { emitDiagnostic } from '../utils/diagnostics.ts';
 import { sleep } from '../utils/timeouts.ts';
 import {
   captureAndroidFreshnessRecoveredAttempt,
@@ -25,8 +24,8 @@ import {
   haveIdenticalDiscriminatingSurfaces,
   markPendingInteractionOutcome,
   retryPendingInteractionOutcome,
-  type InteractionSurfaceSignature,
 } from './interaction-outcome-policy.ts';
+import { runPostGestureStabilityLoop } from './post-gesture-stability.ts';
 import type { SessionState } from './types.ts';
 
 /**
@@ -48,25 +47,6 @@ import type { SessionState } from './types.ts';
  */
 
 const INTERACTION_CHANGE_RECHECK_DELAY_MS = 500;
-const STABILIZATION_DEADLINE_MS = 1_500;
-const STABILIZATION_INTERVAL_MS = 200;
-const STABILIZATION_MIN_ATTEMPTS = 2;
-
-/**
- * Defect 2 (#1542): a bounded extra budget used ONLY when a quiet AX-signature
- * match (two consecutive polls agree) still equals the pre-gesture baseline on
- * the Apple synthesized-gesture lane (see `requiresPostGestureBaselineDistrust`).
- * XCTest's AX tree isn't proactively resynced by a synthesized touch, so it
- * can serve a stale-but-internally-consistent read that two polls agree on
- * without the screen having moved.
- *
- * 2s of real margin over both the poll interval (200ms) and the normal
- * deadline (1.5s) — a near-zero margin between a poll interval and a quiet
- * window is a proven flake source in this codebase (see
- * settle-zero-margin-flake, a week-long contention-flake root cause), so this
- * cap is sized to never come close to that trap.
- */
-const STABILIZATION_DISTRUST_DEADLINE_MS = STABILIZATION_DEADLINE_MS + 2_000;
 
 /**
  * Mutation-side marking, called once per mutating dispatch after the device op
@@ -313,103 +293,19 @@ async function capturePostActionSnapshotAttempt(
   return await params.capture();
 }
 
-export type PostGestureStabilityVerdict = 'trust' | 'distrust' | 'accept-stale';
-
-/**
- * Pure decision at the heart of defect 2's fix (#1542). Called only once a
- * quiet AX-signature match has already been observed (two consecutive
- * post-gesture polls agree); decides whether that agreement is trustworthy
- * "settled" evidence or a stale-but-consistent AX read that happens to still
- * equal the pre-gesture baseline.
- *
- * - `trust`: accept immediately — the platform doesn't need baseline distrust
- *   (Android is fresh by construction), there is no usable baseline, or the
- *   quiet signature genuinely differs from the pre-gesture baseline.
- * - `distrust`: the quiet signature still equals the baseline AND the bounded
- *   distrust cap has not expired — keep polling, do not accept as final.
- * - `accept-stale`: the cap expired and the signature still equals the
- *   baseline. A genuinely inert gesture (e.g. scroll already at an edge) is
- *   the honest read at this point, so it is accepted — but flagged, so a
- *   stale-accept is distinguishable from an ordinary settle in diagnostics.
- *
- * The baseline comparison is `classifyBaselineSurfaceEvidence` — see its doc
- * for why it is subset-tolerant set membership over identified content, not
- * whole-array equality and not a boolean (#1542/#1569, both live-verified).
- * `'ambiguous'` (no comparable evidence) falls through to `trust`, same as
- * `'changed'`.
- */
-export function decidePostGestureStabilityVerdict(params: {
-  needsBaselineDistrust: boolean;
-  baselineSignature: InteractionSurfaceSignature | undefined;
-  quietSignature: InteractionSurfaceSignature;
-  elapsedMs: number;
-  distrustCapMs: number;
-}): PostGestureStabilityVerdict {
-  const { needsBaselineDistrust, baselineSignature, quietSignature, elapsedMs, distrustCapMs } =
-    params;
-  if (!needsBaselineDistrust || !baselineSignature?.length) return 'trust';
-  if (classifyBaselineSurfaceEvidence(baselineSignature, quietSignature) !== 'unchanged') {
-    return 'trust';
-  }
-  return elapsedMs < distrustCapMs ? 'distrust' : 'accept-stale';
-}
-
-type CapturedSurface<T> = {
-  value: T;
-  signature: InteractionSurfaceSignature;
-  backend: string | undefined;
-};
-
-async function captureInteractionSurface<T>(
-  capture: () => Promise<T>,
-  readSnapshot: (result: T) => SnapshotState,
-  initial?: T,
-): Promise<CapturedSurface<T>> {
-  const value = initial ?? (await capture());
-  const snapshot = readSnapshot(value);
-  return {
-    value,
-    signature: buildInteractionSurfaceSignature(snapshot.nodes),
-    backend: snapshot.snapshotQuality?.backend,
-  };
-}
-
-function emitPostGestureSettleDiagnostic(
-  verdict: 'trust' | 'accept-stale',
-  action: string,
-  attempts: number,
-  durationMs: number,
-): void {
-  if (verdict === 'accept-stale') {
-    emitDiagnostic({
-      level: 'warn',
-      phase: 'post_gesture_snapshot_stale_accept',
-      data: { action, attempts, durationMs, matchedPreGestureBaseline: true },
-    });
-    return;
-  }
-  emitDiagnostic({
-    level: attempts > 2 ? 'info' : 'debug',
-    phase: 'post_gesture_snapshot_stabilized',
-    data: { action, attempts, durationMs },
-  });
-}
-
 export type PostGestureStabilizedResult<T> = {
   value: T;
-  /**
-   * Present ONLY when the accept-stale verdict is corroborated by full-surface
-   * evidence: every discriminating entry of the quiet capture matches the
-   * pre-gesture baseline exactly, in both directions
-   * (`haveIdenticalDiscriminatingSurfaces`). The bare verdict is NOT enough —
-   * it is subset-tolerant by design, and a successful scroll that replaced
-   * every list cell under fixed chrome still reads accept-stale (#1601 review
-   * P1). Callers surface this to the agent: a diagnostics-only signal let one
-   * benchmark run burn 40 calls re-issuing scrolls that moved nothing (#1600).
-   */
+  /** See `PostGestureStabilityOutcome` in post-gesture-stability.ts (#1600/#1601). */
   gestureNoEffect?: { action: string; positionals: string[] };
 };
 
+/**
+ * Session-aware adapter over the pure stability loop
+ * (`post-gesture-stability.ts`): reads the pending record, supplies the
+ * interaction-surface comparators from interaction-outcome-policy as hooks,
+ * and — as the R7 owner — clears `postGestureStabilization` once the loop
+ * returns, on settle and timeout alike.
+ */
 export async function capturePostGestureStabilizedResult<T>(params: {
   session: SessionState | undefined;
   capture: () => Promise<T>;
@@ -422,92 +318,31 @@ export async function capturePostGestureStabilizedResult<T>(params: {
     return { value: params.initial ?? (await capture()) };
   }
 
-  const needsBaselineDistrust = requiresPostGestureBaselineDistrust(session.device);
-  const startedAt = Date.now();
-  let attempts = 1;
-  let previous = await captureInteractionSurface(capture, readSnapshot, params.initial);
-  let baselineSignature = pending.baselineSignature;
-  let baselineBackend = pending.baselineBackend;
-  // Extended past STABILIZATION_DEADLINE_MS only when the distrust verdict
-  // fires below; the ordinary (non-distrust) timeout path is unaffected.
-  let effectiveDeadlineMs = STABILIZATION_DEADLINE_MS;
-
-  while (attempts < STABILIZATION_MIN_ATTEMPTS || Date.now() - startedAt < effectiveDeadlineMs) {
-    await sleep(STABILIZATION_INTERVAL_MS);
-    attempts += 1;
-    const current = await captureInteractionSurface(capture, readSnapshot);
-    if (areInteractionSurfaceSignaturesStable(previous.signature, current.signature)) {
-      const elapsedMs = Date.now() - startedAt;
-      // A capture plan may fall back or be pre-empted by the XCTest-channel
-      // penalty at any time, so the backend can change mid-poll. Backends do
-      // not agree on which nodes exist, so this pair says nothing about the
-      // gesture: adopt it as the baseline and keep going rather than concluding
-      // from it (#1569).
-      if (baselineSignature && baselineBackend !== current.backend) {
-        emitDiagnostic({
-          level: 'debug',
-          phase: 'post_gesture_snapshot_baseline_rebased',
-          data: { action: pending.action, from: baselineBackend, to: current.backend, attempts },
-        });
-        baselineSignature = current.signature;
-        baselineBackend = current.backend;
-        previous = current;
-        continue;
-      }
-      const verdict = decidePostGestureStabilityVerdict({
-        needsBaselineDistrust,
-        baselineSignature,
-        quietSignature: current.signature,
-        elapsedMs,
-        distrustCapMs: STABILIZATION_DISTRUST_DEADLINE_MS,
-      });
-      if (verdict === 'distrust') {
-        effectiveDeadlineMs = STABILIZATION_DISTRUST_DEADLINE_MS;
-        previous = current;
-        continue;
-      }
-      clearPostGestureStabilization(session);
-      emitPostGestureSettleDiagnostic(verdict, pending.action, attempts, elapsedMs);
-      return buildAcceptedStabilizedResult(verdict, pending, current);
-    }
-    previous = current;
-  }
-
-  clearPostGestureStabilization(session);
-  emitDiagnostic({
-    level: 'warn',
-    phase: 'post_gesture_snapshot_stabilization_timeout',
-    data: {
-      action: pending.action,
-      attempts,
-      durationMs: Date.now() - startedAt,
-    },
-  });
-  return { value: previous.value };
-}
-
-/**
- * The no-effect claim needs BOTH the accept-stale verdict AND full-surface
- * corroboration (`haveIdenticalDiscriminatingSurfaces`): the verdict alone is
- * subset-tolerant, and a successful scroll that replaced every list cell
- * under fixed chrome still reads accept-stale (#1601 review P1).
- */
-function buildAcceptedStabilizedResult<T>(
-  verdict: 'trust' | 'accept-stale',
-  pending: NonNullable<SessionState['postGestureStabilization']>,
-  current: CapturedSurface<T>,
-): PostGestureStabilizedResult<T> {
-  const corroborated =
-    verdict === 'accept-stale' &&
-    haveIdenticalDiscriminatingSurfaces(pending.baselineSignature ?? [], current.signature);
-  if (!corroborated) return { value: current.value };
-  return {
-    value: current.value,
-    gestureNoEffect: {
+  const outcome = await runPostGestureStabilityLoop({
+    pending: {
       action: pending.action,
       positionals: pending.positionals ?? [],
+      baselineSignature: pending.baselineSignature,
+      baselineBackend: pending.baselineBackend,
     },
-  };
+    needsBaselineDistrust: requiresPostGestureBaselineDistrust(session.device),
+    initial: params.initial,
+    hooks: {
+      capture,
+      readSurface: (value) => {
+        const snapshot = readSnapshot(value);
+        return {
+          signature: buildInteractionSurfaceSignature(snapshot.nodes),
+          backend: snapshot.snapshotQuality?.backend,
+        };
+      },
+      signaturesStable: areInteractionSurfaceSignaturesStable,
+      classifyBaselineEvidence: classifyBaselineSurfaceEvidence,
+      surfacesIdentical: haveIdenticalDiscriminatingSurfaces,
+    },
+  });
+  clearPostGestureStabilization(session);
+  return outcome;
 }
 
 function isPostGestureStabilizingAction(
